requirements.txt
langchain-groq
langchain-community
langgraph
streamlit
duckduckgo-search
python-dotenv